✨ feat(services.py): add send_answer function to handle message replies and translations 🧹 chore(main.py): refactor imports to use send_answer from services instead of utils 🧹 chore(summary.py): update import to use check_quota from services 🧹 chore(utils.py): remove send_answer and check_quota functions to clean up code and avoid duplication
